refactor: LoginView 삭제 후 LoginViewController에서 View 구현, 카카오&페이스북 로그인 구현
